[MIRROR] [MDB IGNORE] Add "Hall of Fame" emergency shuttle (#7076)
Original PR: https://github.com/tgstation/tgstation/pull/77074
-----
## About The Pull Request
This adds a new emergency shuttle called the **Hall of Fame**.

![2023-07-28 06 49
58](https://github.com/tgstation/tgstation/assets/5195984/d818058c-f031-472a-8314-8e4c9054c807)

It's designed around persistence. The goal is to have the shuttle store
memories, photos, and trophies for the crew to see!

## Why It's Good For The Game
Cool way for the crew to store and share memories.

## Changelog
:cl:
add: Add a new 'Hall of Fame' emergency shuttle. It even comes with it's
own nifty photo album.
/:cl:

---------

Co-authored-by: Tim <timothymtorres@gmail.com>