[MIRROR] [MDB IGNORE] Fix TTS leaking identity of `TRAIT_UNKNOWN` mobs (#7091)
Original PR: https://github.com/tgstation/tgstation/pull/77279
-----
## About The Pull Request

If you have `TRAIT_UNKNOWN`, your chat messages will not TTS. 

Maybe a more fun alternative is to pick a generic voice for
`TRAIT_UNKNOWN` mobs, with a filter that makes the voice sound obscured
or something, idk. This is easy though.

## Why It's Good For The Game

`TRAIT_UNKNOWN` is intended to leave no hint you are who you are, so to
suddenly speak in a very distinguishable voice that becomes useless
quick.

This means someone paying attention can easily pinpoint who the traitor
behind the suit is, or the heretic behind the cloak.

(It also probably means that this screws up for people wearing a
disguise (+voice changer) or lings. Should visit that later)

## Changelog

:cl: Melbert
fix: Traitors using the sneak suit and heretics under the effects of
cloak no longer leak their identity via Text to Speech
/:cl:

---------

Co-authored-by: MrMelbert <51863163+MrMelbert@users.noreply.github.com>