[MIRROR] [MDB IGNORE] Restores some mechs being able to pivot step (move and turn in the same keypress) and diagonal movement (#7141)
Original PR: https://github.com/tgstation/tgstation/pull/77345
-----

## About The Pull Request

Most mechs couldn't move diagonally. This restores that for all but the
Clarke (which is a vehicle on treads)

The Clarke and Odysseus once again have pivot step, lost during the
refactor for unknown reasons. It was literally just right there, it was
an easy fix. Clarkes once again move like a tank!

## Why It's Good For The Game

So the way that mech movement works, by blocking diagonal movement, what
instead happens is that the mech eats the move command and goes on
movement cooldown. By pressing two buttons rather than one, you come to
a full stop and no longer can move. I cannot think of a more hostile
player-facing feature than this. It utterly gutted the usability of
mechs for no real apparent benefit. As such, I'm just removing it on all
but the most relevant mech, the Clarke. This should make it much more
approachable to pilot mechs and not feel like the game is malfunctioning
half the time.

Clarkes and Odysseus mechs had a really cool feature. They turned and
moved in the same button press. This made them very nimble mechs and fun
to utilize. They lost this feature in the refactor for, once again, not
much apparent reason since this was literally a line change to restore.

## Changelog
:cl:
fix: Restores a lost feature, Clarkes and Odysseus mechs once again can
pivot on a dime and step in the same button press.
qol: Mechs universally have diagonal movement as to avoid unresponsive
movement, as block diagonal movement was having unintended side effects
and was actually intended for pivot step mechs.
/:cl:

---------

Co-authored-by: necromanceranne <40847847+necromanceranne@users.noreply.github.com>