[MIRROR] [MDB IGNORE] Fixes Cell Message With Mulebots (#7158)
Original PR: https://github.com/tgstation/tgstation/pull/77374
-----

## About The Pull Request
Fixes the message when removing/adding a cell to mulebots using the
mulebot as the user, resulting in sentient mulebots receiving the
message "You add/pry out the cell in the mulebot" despite not being the
one adding/removing the cell. First time making a pr for this codebase,
so let me know if I fucked something up.
## Why It's Good For The Game
The mulebot isn't moving the cell, whoever IS moving it is. Also bug
bad.
## Changelog
:cl:
fix: Fix mulebots receiving the wrong message when thier cell is
added/removed.
/:cl:

---------

Co-authored-by: Wisemonster <87689371+Wisemonster@users.noreply.github.com>
Co-authored-by: san7890 <the@san7890.com>