Adds an Indebted negative quirk to the game. (#77138)
## About The Pull Request
This Pull Request brings a new negative quirk to the list of negative
quirks: Indebted. As the title suggests, you start the round with a
pretty high debt of roughly 15.000 credits (give or take up to 1.250
credits), and everytime the account balance would increment, 75% of
what's earned will be pushed toward solving it instead.

Unlike other quirks, it is hidding from medical HUDs and the health
analyzers, because I frankly believe it's quite a stretch putting it on
the same plane as other quirks which are of a more "psycho-physical"
nature. However, examining the medical records will return a remark on
how the quirk holder has had some difficulties paying the checkup bill,
hinting it greatly. Examining their ID accurately will also report the
entity of the debt.

Ok, the PR is now ready. It fancies an achievement and a pin too, for
those who actually extinguish the debt.

## Why It's Good For The Game
I think it's cool to have a quirk that affects players from an angle
different than most of the others. It doesn't affect the mob directly,
brings no positive or negative moodlet, gives items, but it does cripple
what they can do with their own bank account. Sure, it can be
circumverted by "borrowing" another player mob's ID card, but frankly it
has a mild negative value on par with "Family Heirloom" so it's not a
big deal.

## Changelog

:cl:
add: Added an Indebted negative quirk to the game, which gives the
holder's bank account a debt averaging at 15000 credits (with a
variation of 1250 cr) and forces 75% of all earnings towards solving it.
A little prize awaits those who actually extinguish it.
/:cl: