[MIRROR] [MDB IGNORE] Alert a Discord role when a new player joins (#7203)
Original PR: https://github.com/tgstation/tgstation/pull/77432
-----
## About Why It's Good For The Game The Pull Request
What it says on the tin. A requested feature for people who like to keep
an eye on swaths of new players. A config that is off by default, but a
Discord role ID can be added to have TGS ping that role.
Also sends alerts to Discord if a player with sketchy telemetry joins.

## Changelog
:cl: Tattle
admin: Pings an admin role on discord when a new player joins (when
enabled by config)
/:cl:

---------

Co-authored-by: tattle <66640614+dragomagol@users.noreply.github.com>
Co-authored-by: tattle <article.disaster@gmail.com>