[MIRROR] [MDB IGNORE] Fixes gibtonite explosions producing flames and flashes (#7217)
Original PR: https://github.com/tgstation/tgstation/pull/77344
-----
## About The Pull Request
On the tin. Was introduced in #77118 and uhh yeah

## Why It's Good For The Game

Restoring previous behavior.
To be quite honest I can live with the flames, as they mostly aren't
that much of an issue for a competent miner. Avoiding them or bringing
an extinguisher with you is easy enough; if anything, these flames may
or may not cook loose ore and turn it into sheets, which isn't favorable
for the miner. But for consistency's sake it's better if we restore it
to the previous effects.
![fikou shield
activate](https://github.com/tgstation/tgstation/assets/75863639/1e1f7869-5832-44ad-8778-885fd60695d5)


## Changelog



:cl:
fix: fixed gibtonite explosions having flames and flash ranges
/:cl:

---------

Co-authored-by: Sealed101 <cool.bullseye@yandex.ru>