[MIRROR] [MDB IGNORE] Polishing the greyscale modify menu's lackluster support for non-atom targets. (#7218)
Original PR: https://github.com/tgstation/tgstation/pull/77322
-----
## About The Pull Request
So, I've been recently told that Skyrat uses the greyscale modify menu
for loadouts, and the new ui state kinda borked it. I honestly haven't
taken the possibility that the target could be anything but a subtype of
`/atom` (and still work) into account because much of the code assumes
the target is an atom. It's kinda crappy. Also I hadn't noticed we've an
`unlocked` variable, which makes `vv_mode` var superflous, so i'm going
to remove the latter.

## Why It's Good For The Game
Code improvement, also fixing a downstream issue.

## Changelog
N/A

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>