Makes the ectoplasm anomaly not randomly move so deadchat can actually control it (#77598)
## About The Pull Request

ectoplasmic anomalies randomly move. this makes it so trying to control
it as deadchat is like wrangling a fish. it's probably an oversight more
than anything as most other deadchat controlled things don't move on
their own for no reason
## Why It's Good For The Game

this is probably an oversight more than anything. ectoplasm anomalies
are supposed to be deadchat controllable but they move so damn often
that really isn't feasible. discovered when deadchat was wondering why
the anomaly was just wandering off into nowhere
## Changelog
:cl:
fix: Makes the ectoplasm anomaly not randomly move so deadchat can
actually control it
/:cl:

Co-authored-by: oilysnake <63020759+oilysnake@users.noreply.github.com>