[MIRROR] [MDB IGNORE] Reflavors the Mosin to be a surplus rifle from the past IC 200 years, rather than from 670 years ago in game. Allergy warning: May contain microscopic silverscale buff (#7332)
Original PR: https://github.com/tgstation/tgstation/pull/77169
-----
## About The Pull Request

As the beautiful title may imply:

The Mosin-Nagant rifle has been reflavored into something _slightly_
more modern for the setting (The game currently takes place in 2563 you
know). Rather than a gun made in 1891 ending up in the hands of cargo
techs 670 years in the future, instead aspiring revolutionaries,
larpers, and eastern europeans in general will be seen arming themselves
with something at least a few hundred years more modern.

<details>
<summary>The Sakhno Precision Rifle, the new Nagant</summary>
<br>


![image](https://github.com/tgstation/tgstation/assets/82386923/d26d0bd6-344e-40a2-879c-303b948ce9bd)

"A Sakhno Precision Rifle, a bolt action weapon that was (and certainly
still is) popular with frontiersmen, cargo runners, private security
forces, explorers, and other unsavoury types. This particular pattern of
the rifle dates back all the way to 2440."

</details>

<details>
<summary>The Sakhno M2442 Army, surplus power</summary>
<br>


![image](https://github.com/tgstation/tgstation/assets/82386923/eabcd880-3b0e-4582-a306-000b2e46f4af)

"A modification of the Sakhno Precision Rifle, _Sakhno M2442 Army_ is
stamped into the side. It is unknown what army this pattern of rifle was
made for or if it was ever even used by an army of any sort. What you
can discern, however, is that its previous owner did not treat the
weapon well. For some reason, there's moisture all through the
internals."

</details>

<details>
<summary>The Sakhno - Zhihao Sporting Rifle, weapons for the
rich</summary>
<br>


![image](https://github.com/tgstation/tgstation/assets/82386923/f7e2b6d8-0346-4376-9673-43e405b88a17)

"An upgrade and modernisation of the original Sakhno rifle, made with
such wonders as modern materials, a scope, and other impressive
technological advancements that, to be honest, were already around when
the original weapon was designed. Surprisingly for a rifle of this type,
the scope actually has magnification, rather than being decorative."

</details>

<details>
<summary>The Lionhunter and Enchanted Bolt Action</summary>
<br>


![image](https://github.com/tgstation/tgstation/assets/82386923/1557cd36-f7ca-4fb9-b1a2-49f07c24af0a)

Because it wouldn't make sense for heretics and wizards to be summoning
brand new military green bolt action guns, some wood-like and
significantly fancier looking variations of the sprite now exist for
both of the guns.

</details>



</details>

<details>
<summary>How these look in game:</summary>
<br>


![image](https://github.com/tgstation/tgstation/assets/82386923/aaed7e56-4c0c-406c-bfc9-6188b9231f26)


</details>

## Why It's Good For The Game

<details>
<summary>See a before and after:</summary>
<br>

The sprites on the left of the arrow are what they are now, and the
sprites on the right are what they will turn into, note that there's one
old sprite to several new sprites, this is because guns that were
represented by the same sprite before now have their own unique looks.

These are subject to be slightly different than what you see here, but
will be more or less like these images.


![image](https://github.com/tgstation/tgstation/assets/82386923/73acecb5-e16c-43c2-b690-e1240e3cd06c)


![image](https://github.com/tgstation/tgstation/assets/82386923/188f6de3-1f89-4c26-9e12-2c5a3bffcdbf)


![image](https://github.com/tgstation/tgstation/assets/82386923/c3f7a4da-f4b3-4864-8bf6-314c84ac528a)


![image](https://github.com/tgstation/tgstation/assets/82386923/5a64327c-de63-44cf-9b02-7e90d80a753a)

</details>

While I understand the appeal of wanting a shitty, old rifle to have in
the game for everyone's surplus needs, the Mosin (and a good few other
guns too, looking at you 1911) is just a fair bit too old in my book.
About 670 years old. I find it more believable that while there would
still be some shitty surplus weapon that cargo techs and larpers would
non-stop fawn over, it'd be something that's old by the standards of the
year 2563 where the game takes place, vs old by the standards of 2023.

I think it'd greatly help the sense of the fact that we exist many
hundreds of years in the future if we stop using guns made before the
first world war.

## Changelog
:cl:
image: The Mosin-Nagant has been given new sprites and a reflavor,
looking for the old rifle? Look for the Sakhno Precision Rifle.
balance: The tiniest balance thing, but since Silverscales use the
Sakhno-Zhihao rifle, which has a scope on it, their main weapon now has
a scope.
sound: The cargo rifle now has a new, considerably more rifle sized
firing sound. Gotten from tgmc from
https://github.com/tgstation/TerraGov-Marine-Corps/pull/12280.
/:cl:

---------

Co-authored-by: Paxilmaniac <82386923+Paxilmaniac@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>