[MIRROR] [MDB IGNORE] Fixes an issue with eyes not properly storing the old_eye_color var (#7493)
Original PR: https://github.com/tgstation/tgstation/pull/77692
-----
## About The Pull Request

Somewhere down the line with all the refactors the code that saved the
previous eyes' color stopped working.

This just fixes it. It's used for things such as qol for the high
luminosity eyes. It can be considered a fix and QoL alike but I'm going
to label it as QoL.


![image](https://github.com/tgstation/tgstation/assets/13398309/74645c3a-92de-4ae0-86e4-f0d011549da0)

## Why It's Good For The Game

Restores previous functionality that was lost in a refactor

## Changelog

:cl:
qol: inserting new high luminosity eyes will now properly remember the
previous eyes' color and default to that initially
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>