[MIRROR] [MDB IGNORE] Fixes some heretic bugs (?) (#7543)
Original PR: https://github.com/tgstation/tgstation/pull/77611
-----
## About The Pull Request

Items worn on a void cloak's suit storage are no longer visible on the
stripping menu.

Heretic influences don't leave scannable fingerprints anymore.

## Why It's Good For The Game

> Items worn on a void cloak's suit storage are no longer visible on the
stripping menu.

Described as a bug by Melbert. Shouldn't be able to bypass the sneaky
item with an oversight. This is done by giving any equipped item in the
suit storage slot the no_strip trait with the cloak.

> Heretic influences don't leave scannable fingerprints anymore.

Likely a lot more controversial, but this is bad for a lot of reasons.

1. It screws over new heretic players with some turbonerd knowledge that
they had no hope of figuring out.
Who in their right mind would assume that influences, literal rifts in
reality, somehow have fingerprints in their non-existent surface that
some detective could walk up and scan, despite that same rift being able
to tear your arms off if you poke it too much?

2. This is literally just bypassed by using a damp rag on the rift.

This means that this is a pure knowledge check, trivial to step over if
you know it, impossible to figure out and potentially round-ending if
you don't.

3. It's too easy!

Detectives shouldn't scan the influence directly for prints, they should
scan the doors, check the prints and fibers, interview people whose
fibers and/or prints match up... not get a 'Yep, Joe's A Heretic' huge
blaring confirmation. It's not detective forensic analysis, it's just
using jank on jank to confirm without a doubt the identity of a hapless
antagonist. I'm all for emergent gameplay, but this isn't emergent
gameplay, it's an emergent unfair knowledge check.

4. It appears to orgiinate from a bug?

It has a flag that makes it so that prints aren't left behind by
attack_handing it, but the prints still show up? I changed it to the one
that just makes fingerprints impossible to appear at all, so that should
clear things up.

And yes, Ided from these thins

Untested, it's 3 am give me a break

## Changelog

:cl:
fix: Items worn on a void cloak's suit storage are no longer visible on
the stripping menu.
qol: Misclicking the turf below a heretic influence no longer leaves a
fingerprint.
/:cl:

---------

Co-authored-by: carlarctg <53100513+carlarctg@users.noreply.github.com>