[MIRROR] [MDB IGNORE] Adds Blood-drunk and demonic frost miner boss music. (#7649)
Original PR: https://github.com/tgstation/tgstation/pull/78123
-----
## About The Pull Request

Acts as a continuation of PR #77149 for boss music functionality and
implements a BDM and demonic frost miner boss music theme.

## Why It's Good For The Game

More music is good, but I do have some gripes with my own PR. This
particular track relies on instrumentation that when compressed just
doesn't sound as good, and the in-game version is noticeably less
enjoyable that the high quality version. I wish I could help the track
out more, but as is it's already at 811 kb which is barely in line with
file requirements, so i just can't justify bloating the audio file sizes
to make it sound better. You notice this kind of problem a lot with the
higher runtime music and background tracks. It just feels a bit more
clunky than hierophant, but what are you gonna do right?

Here's a track sample. I had to use a third-party thingy to convert it
into mp4 so it might sound off but whatever. Thanks!


https://github.com/tgstation/tgstation/assets/77305289/d9f64665-ff9c-4ba4-95d1-6ad7d408f868


## Changelog

sound: adds BDM/frost miner theme

/:cl:

---------

Co-authored-by: RICK IM RI <77305289+tommysalami3@users.noreply.github.com>