[MIRROR] [MDB IGNORE] Replace DreamAnnotate action with a python script (#7701)
Original PR: https://github.com/tgstation/tgstation/pull/78225
-----

## About The Pull Request

This PR removes the "Annotate Lints" job step and merges it with the
"Run Linters" step above. To achieve this, I wrote a python script that
should be identical to the action. I even added the ability to read the
output from a file to the script if we ever needed that, but I decided
to have the job step pipe the output into the script instead.
## Why It's Good For The Game

It always bugged me a bit that we had to check the results for a
separate step if we wanted to see the linter results for dm code. I also
noticed a few people getting confused as to why their CI failed on
linters.

Turns out that the action is just a few lines that match the
dreamchecker output and reformat it to a format that GitHub can annotate
code with. It's so brain dead simple that it shouldn't need to be a
whole new step, and for the previous two reasons.
## Changelog
not playerfacing

---------

Co-authored-by: distributivgesetz <distributivgesetz93@gmail.com>