[MIRROR] [MDB IGNORE] Fixes a fruit wine runtime and fixes runtime when trying to add inedible plant seeds to the seed extractor (#7756)
Original PR: https://github.com/tgstation/tgstation/pull/78319
-----
Fixes https://github.com/tgstation/tgstation/issues/78315
Fixes https://github.com/tgstation/tgstation/issues/69282 (same issue)
Fixes https://github.com/Skyrat-SS13/Skyrat-tg/issues/23586 (same issue)

## About The Pull Request

Jacquerel end me please.

This fixes the linked issue and the other issue that was mentioned in
the comments. They are unrelated, but this is a two for one.


First bug is caused by fruit wine trying to blend `filling_color`, which
is actually unset for the majority of plants. As such it will runtime
when trying to make wine from just about every plant in the game. It
could be a nice project for someone to set those for all the plants so
they have more accurate colors when smashed or made into fruit wine...

But for now the bug is fixed by just having it use the reagent's `color`
for the blending when the `filling_color` var is unset.

---

~~As for the other issue mentioned in the comments of #78315, this is
because it turns out the code in hydroponics.dm treats `/obj/item/grown`
and `/obj/item/food/grown` the same, so we need to make sure that their
constructors match (even if we aren't using the args).~~

~~It's an issue born ultimately of bad OOP design. Both should be
basetype of obj/item/grown really but the way that food code works it
has to be this way.~~

As per Jacquerel 's request I am going to remove this portion of the
fix. I've fixed the runtime that occurs when adding the seeds to the
extractor but as for the fix for the mutation traits I will leave that
in their capable hands.


## Why It's Good For The Game
## Changelog

:cl:
fix: fixes a bug that would cause grown inedible plant seeds (like tower
cap) to vanish from existence upon being added to the seed extractor
fix: fixes a issue that would cause fruit wine to bug out when trying to
blend its reagent color

/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>