[MIRROR] [MDB IGNORE] Adds explodable walls for space ruins (#7757)
Original PR: https://github.com/tgstation/tgstation/pull/78239
-----

## About The Pull Request

Adds an explodable subtype of indestructible walls. Retains most forms
of indestructibility, but a nearby explosion will break it down.

Added an example ruin with a buried ~~supply~~ costume cache to
demonstrate the walls. All the tools to blow it up are concealed around
the ruin.


![image](https://github.com/tgstation/tgstation/assets/5479091/3ec663b8-4bea-462c-8c0e-59739bc42d50)

## Why It's Good For The Game

More options for space ruin puzzles are always good to have.

## Changelog
:cl:
add: There's a new space ruin in town, be on the lookout for a hidden
supply cache.
add: Added a new type of wall which can only be destroyed by blowing it
up.
/:cl:

---------

Co-authored-by: Thunder12345 <Thunder12345@users.noreply.github.com>