Unholy and Eldritch water are self-consuming (#78408)
## About The Pull Request

Unholy and Eldritch water are self-consuming! They don't need a liver to
be processed.

## Why It's Good For The Game

Consistency with holy water, is reasonable and intuitive if you know the
former was already self consuming. To keep it like this is inconsistent.

## Changelog

:cl:
balance: Unholy and Eldritch water are self-consuming like holy water!
They don't need a liver to be processed.
/:cl: