[MIRROR] [MDB IGNORE] Fixes tackling issues with stun immune tacklers (#7894)
Original PR: https://github.com/tgstation/tgstation/pull/78494
-----

## About The Pull Request

When a human uses gripper gloves and such to perform a tackle, it does
some stuff with knockdowns. First, at the start of the tackle, the
tackler is given 1 or 2 seconds of knockdown. This knockdown bypasses
stun immunity. If the tackle and its rolled to be an expert or monster
tackle, then this initial knockdown is cleared. However, the knockdown
clearing does NOT bypass stun immunity. This ironically meant that stun
immune tacklers would remain knocked down when landing a monster tackle,
and it would cause a runtime because knocked down people can't grab.
This PR just makes the knockdown clear bypass stun immunity as well.
Closes #78441.
## Why It's Good For The Game

Bugfixes are good.
## Changelog
:cl:
fix: Stun immune people should no longer have issues with gripper gloves
and other tackle gloves.
/:cl:

---------

Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>