[MIRROR] [MDB IGNORE] Planetary atmos fix (#7915)
Original PR: https://github.com/tgstation/tgstation/pull/78602
-----

## About The Pull Request

The code for generating random atmospheres for lavaland and icemoon has
variables that define the minimum and maximum pressure that the
atmosphere can be. However, due to an oversight, it was possible for the
actual pressure to be below this minimum. This caused problems for the
flight potion wings, as the wings will not work if the pressure is lower
than the minimum lavaland/icemoon pressure.

I tested this by making a quick test proc that would generate lavaland
atmos 100 times, and count how often the final pressure was lower than
the minimum pressure. I ran it a few times after making my changes, and
it returned 0 failures every time, so I'm confident that my code works.
## Why It's Good For The Game

I think its good practice for minimums and maximums variables like this
to actually be hard limits. Furthermore, if there's _one_ place that
mining loot should actually be useful, then its outside where you mine.
## Changelog
:cl:
fix: The flight potion wings will no longer fail to work on
lavaland/icemoon on rare occasions.
/:cl:

---------

Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>