[MIRROR] [MDB IGNORE] Prevent things from the feature list generating on turfs with flora (#7919)
Original PR: https://github.com/tgstation/tgstation/pull/78558
-----
## About The Pull Request

which basically means no geysers on the same turf as a plant turf.
they're both structures but the for loop before this one checks for
same-type structures in a radius. and bush geysers look stupid! and
weird! and stupid!

## Changelog

:cl:
fix: fixed geysers spawning on turfs with plants
/:cl:

---------

Co-authored-by: Sealed101 <cool.bullseye@yandex.ru>