[MIRROR] [MDB IGNORE] You can now pick the limb you want replaced with the Prosthetic Limb Quirk. (#7929)
Original PR: https://github.com/tgstation/tgstation/pull/78517
-----

## About The Pull Request

The Prosthetic Limb quirk allows you to pick the limb you want. The
default is still random
## Why It's Good For The Game

Good for character customization, which is always nice. Statics who got
a specific limb replaced can now represent that properly in-game.
Also, this makes the quirk show up on the dummy on the character
customization again. This *does* cause a slight issue where when you
switch to another character, it keeps the previous' skin color, but it's
minor enough that i think the benefits are worth the drawback. If asked
i will make it not appear on the dummy though.
## Changelog
DATA_, with great help from Kapu:cl:
add: You can now choose the prosthetic you want with the Prosthetic Limb
quirk.
/:cl:

---------

Co-authored-by: DATAxPUNGED <44149906+DATA-xPUNGED@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>