[MIRROR] [MDB IGNORE] Sends a message to admins when a headrev/head disconnects during a revolution. They are no longer immediately treated as dead but are given a chance to reconnect (#7943)
Original PR: https://github.com/tgstation/tgstation/pull/78586
-----

## About The Pull Request
As the title says, headrevs and heads of staff during a revolution won't
be disqualified when they either disconnect or go AFK. Instead, they're
given a 2 minute timer to reconnect or stop being AFK and a warning is
sent out to admins so that they can offer up the player if time is close
to running out.
Two admin messages are sent, one when they're initially disconnected or
AFK and one sent a minute after, giving the admin a minute to act on the
log.

## Why It's Good For The Game
Disconnects can happen for any number of reasons, but sometimes they can
be unintentional, like the game crashing or briefly losing connection.
This will give headrevs and heads more leeway. Additionally, a solo
headrev who disconnects before they've converted anyone could be spotted
by admins and replaced so that the revolution gamemode isn't wasted.

## Changelog
:cl:
balance: Head revolutionaries and heads of staff are no longer
immediately considered disqualified when going AFK or disconnecting and
are given a 2 minute grace period.
admin: Admins now get a log when a head revolutionary or head of staff
disconnects or goes AFK during a revolution. They also get the same log
1 minute after to give them a chance to act on the information.
/:cl:

---------

Co-authored-by: Watermelon914 <37270891+Watermelon914@users.noreply.github.com>
Co-authored-by: Watermelon914 <3052169-Watermelon914@users.noreply.gitlab.com>