[MIRROR] [MDB IGNORE] Turns Robot Customers into Basic Mobs (#7945)
Original PR: https://github.com/tgstation/tgstation/pull/78611
-----

## About The Pull Request

I'm 99% sure these were made with the new ai controller framework BEFORE
basic mobs were a thing, so we don't need to bother with writing new AI
or venue interactions or anything weird like that. Just gotta dust off
some of this code and make sure everything is in line with modern code
standards as we convert this guy to the new hotness.
## Why It's Good For The Game

Honestly, there's no perceivable difference between them now. One nice
thing is that we clean up the code since we don't have to set the vars
to override the simple animal "AI", it is just simply allowed to be
free. Most of the information about the bot lied in the ai_controller we
would assign it or the customer info datum, we just need to clean the
vessel.
## Changelog
:cl:
code: Robot Customers have recently been touched codewise, please report
any bugs or unexpected behavior as there really should not be any.
/:cl:

---------

Co-authored-by: san7890 <the@san7890.com>