[MIRROR] [MDB IGNORE] Adds an Admin Tool for the DNA Infuser (#7946)
Original PR: https://github.com/tgstation/tgstation/pull/78496
-----
## About The Pull Request

Adds some an vv admin tool for DNA Infusions. The tool will
automatically grant all of the relevant organs to the target.


![pic1](https://github.com/tgstation/tgstation/assets/21979502/e34548af-9648-4842-a089-02cffc9989cf)

![pic4](https://github.com/tgstation/tgstation/assets/21979502/1c0aa855-e284-4891-800e-717383425b76)

![pic3](https://github.com/tgstation/tgstation/assets/21979502/c955a35f-5835-4271-88cb-f1ab198cb8df)
## Why It's Good For The Game

It was already possible to do this with the organ manipulation tool, but
that's a lot slower and probably has issues when replacing brains. This
new tool will hopefully streamline the testing of new DNA infuser
entries.
## Changelog
:cl:
admin: There is now a tool to apply a DNA Infuser entry to any human.
/:cl:

---------

Co-authored-by: GPeckman <21979502+GPeckman@users.noreply.github.com>