[MIRROR] [MDB IGNORE] A thorough audit of damage procs and specifically their use in on_mob_life() (with unit tests!) (#8008)
Original PR: https://github.com/tgstation/tgstation/pull/78657
-----
## About The Pull Request

This PR is a full audit of damage procs as I found so many needless
calls to `updatehealth()`, which is a fairly expensive proc. The aim
here is to reduce proc call overhead in hot areas of code as well as
general code maintenance/cleanup.

A secondary goal with this is to reduce cargo cult code, _especially_
with regards to `on_mob_life()`. People tend to learn from example so
let's make sure those examples are correct and deliberate, and not just
something blindly copy pasted from another proc.

---

<details> <summary> An egregious example of some needless updatehealth()
calls. This code causes a costly 5n calls once and then another 5
updatehealth() calls per life tick per mob. </summary>


https://github.com/tgstation/tgstation/blob/fcfc1cedf9770b92561f4d8610e3ce91b4229460/code/modules/reagents/chemistry/reagents/drinks/alcohol_reagents.dm#L1594-L1616

Each updatehealth() call then calls up to 7 additional procs, which each
of those call even more procs further down the chain.

If you got everyone drinking some bastion bourbon I'm pretty sure the
server itself blacks out drunk...maybe that was the intent there by some
disgruntled past coder?

</details>


<details> <summary> After these changes it only gets called once per
on_mob_life, and then once in on_mob_metabolize</summary>


![image](https://github.com/tgstation/tgstation/assets/13398309/824eb759-22d3-4c97-9bb7-a83c7d91ecae)

Also got rid of the for loop insanity in favor of simply passing the
heal amount because what even was that.

</details>

---

## Foreword:

I want to state that **THERE ARE NO BALANCE CHANGES IN THIS PR. This is
pure code-janitoring.**

I have personally tested it both in-game and by unit tests that I've
included, but it would still be good to testmerge given how many changes
there are (apologies for that to whoever reviews this!).

## Damage proc portion (aka adjustBruteLoss, adjustToxLoss,
adjustFireLoss, etc):

- Refactors all of the damage procs to return the net change in damage
that was dealt or healed to the mob. Also made the procs 'smart' in that
if there is no net change, they don't call `updatehealth()` needlessly
(even if `updating_health = TRUE`).

- Adds unit tests that make sure that damage procs are returning the
right values for human and basic mobs. Also checks some edge cases like
biotypes and the `TOXIMMUNE` trait, etc.

- Goes through all the instances of damage procs and updates their
arguments to use keywords over positional args. Also replaces `0` args
with `FALSE`.

- Audits all instances of adjustBruteLoss, adjustToxLoss, etc and makes
sure that they are not calling updatehealth() needlessly. That brings us
to the next part...

## on_mob_life() portion:

<details> <summary> Long version </summary>

So `on_mob_life()` is supposed to return `TRUE` if it needs to cause an
`updatehealth()` proc call. Same thing goes for `on_mob_dead()`,
`overdose_start()`, and `overdose_process()`.

Then, when everything has been done, the return values get tallied
together and `updatehealth()` gets called once if needed.


https://github.com/tgstation/tgstation/blob/fcfc1cedf9770b92561f4d8610e3ce91b4229460/code/modules/reagents/chemistry/holder.dm#L839

That's the idea anyway.

There was however, a lot of confusion and lack of coherence across the
various implementations of this proc. With a good amount of copy pasta
cargo cult coding, such as `on_mob_life()` returning `TRUE` when it
didn't have to be. Or worse, like in the bastion bourbon example just
straight up ignoring that aspect of the code and doing its own thing.

There were also a lot of examples of people thinking they had to
`updatehealth()`when a mob was incapacitated (sleeping, paralyzed, etc),
which is not the case. It's not readily obvious looking at the code but
`update_stat()` (one of the many procs that `updatehealth()` calls) gets
called when the `TRAIT_KNOCKEDOUT` gets added.

<details><summary>Shown here</summary>


https://github.com/tgstation/tgstation/blob/fcfc1cedf9770b92561f4d8610e3ce91b4229460/code/modules/mob/living/init_signals.dm#L69-L78

</details>

So basically every time someone returned TRUE after putting a mob to
sleep they were just doing cargo cult coding because it's simply not
needed.

My goal here is to make it extremely consistent and clear what is going
on to reduce this sort of thing, by empowering people to understand what
is actually happening in these procs and culling any bad examples that
might get copy pasted.


</details>

- I've made a define `UPDATE_MOB_HEALTH` and changed each instance of
`return TRUE` or `. = TRUE` to `return UPDATE_MOB_HEALTH` or `. =
UPDATE_MOB_HEALTH`

- Given we are actually using the return value for the parent of this
proc, I felt that it would be clearest if we start off with `. = ..()`
by default. The amount of exceptions to this where you'd need to 'return
..()' later on in the proc were in the single digits in any given
reagent file, and this makes it clearer that we actually _want_ the
return value from the parent.

- With these things in mind, did an audit of `on_mob_life()`, making
sure that `updatehealth()` is being called only as needed.

- As a consequence of calling the parent of `on_mob_life()` first,
`current_tick` now for all intents and purposes starts at `1` instead of
`0`. This makes more sense when you think about it. It doesn't change
anything functionally, and I've accounted for it in any places of the
code where it might have caused conditions to happen sooner by
incrementing the conditional check by 1. There should be no detectable
difference whatsoever.

## Why It's Good For The Game

Makes reagents code a bit more unified and clear to work with, and the
amount of cargo cult is reduced.
Damage procs returning the actual value of damage dealt/healed is also a
good thing.

<details> <summary> And yes, health still updates like it's supposed to
</summary>


![dreamseeker_nwLIhFf9JT](https://github.com/tgstation/tgstation/assets/13398309/42f03c1a-58b4-4bf6-a208-33b561ad057a)

</details>

## Changelog

:cl:
refactor: fixed many instances of updatehealth() either being called
needlessly or not at all within on_mob_life() and in various other parts
of the code
refactor: damage procs now return useful information--the actual net
change in damage on the mob. added a unit test for this
/:cl:

---------

Co-authored-by: Bloop <13398309+vinylspiders@users.noreply.github.com>
Co-authored-by: Jolly-66 <70232195+Jolly-66@users.noreply.github.com>